Update gems and dependencies
This brings the project inline with what Quke is currently using. The intention of the project is to allow for testing of a vanilla interaction with cucumber to confirm how Quke should be setup or behave.

Hence it's handy to keep the dependencies up to date and inline with Quke.